backend. Refactorizado modelo. Adaptado a nuevo DDL y separadas entidades del modelo en archivos propios
